Set pretty hostname on service start
