+++ debian/further-link.lintian-overrides
@@ -2,3 +2,4 @@ further-link: no-manual-page usr/bin/further-link
 further-link: no-manual-page usr/bin/start-further
 further-link: no-manual-page usr/bin/further-link-bluetooth-pairing
 further-link: no-manual-page usr/bin/further-link-bluetooth-encryption
+further-link: no-manual-page usr/bin/further-link-set-pretty-hostname
