+++ debian/further-link.service
@@ -7,6 +7,7 @@ After=network.target
 Type=simple
 Environment="PYTHONUNBUFFERED=1"
 Environment="PYTHONDONTWRITEBYTECODE=1"
+ExecStartPre=/usr/bin/further-link-set-pretty-hostname
 ExecStart=/usr/bin/further-link
 
 [Install]
