+++ setup.cfg
@@ -60,6 +60,7 @@ console_scripts =
     start-further=further_link.start_further:start_further
     further-link-bluetooth-pairing=further_link.util.bluetooth.pairing:main
     further-link-bluetooth-encryption=further_link.util.bluetooth.encryption:main
+    further-link-set-pretty-hostname=further_link.util.hostname:set_pretty_hostname
 
 [bdist_wheel]
 universal = 1
